requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-restful==0.3.10
flask-marshmallow==1.2.1  
marshmallow==3.21.1       
marshmallow-sqlalchemy==1.0.0
flask-jwt-extended==4.6.0
flask-migrate==4.0.7
psycopg[binary]
python-dotenv==1.0.1
email-validator==2.1.1
pytest==8.1.1
pytest-flask==1.3.0